2022/5.auth/python_user_system/pyproject.toml
[tool.poetry]
name = "python_user_system"
version = "0.1.0"
description = ""
authors = ["yingshaoxo <yingshaoxo@gmail.com>"]
packages = [
    { include = "**/*.py" },
]

[tool.poetry.dependencies]
python = "^3.9"
PyJWT = "^2.3.0"
fastapi = "^0.74.1"
uvicorn = "^0.17.5"
requests = "^2.27.1"
SQLAlchemy = "^1.4.31"
databases = "^0.5.5"
aiosqlite = "^0.17.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
yapf = "^0.32.0"
ipykernel = "^6.9.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
dev = "src.main:start"
test = "src.test.entry:main"